remove domain stuff because of reason I forgot? do it manually?

+++ teardown.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
+if [ "$AWS_PROFILE" != "private" ]; then
+    echo "AWS_PROFILE must be \"private\""
+    exit 1
+fi
+
 terraform -chdir=terraform destroy -auto-approve
